Move service files to core/ subfolder
